change gradient files, remove attribute.h, add trim_start inside common.h
