--- src/LinearGradient.cpp
@@ -1,5 +1,0 @@
-#include "LinearGradient.h"
-
-Paint LinearGradient::to_paint() {
-  return {};
-}
